refactor: extract WorkspaceContextMenuCollectionView and remove stale duplicates
Extracted WorkspaceContextMenuCollectionView into a shared file using
a closure-based workspacesProvider instead of weak back-references to
parent views. Removed stale duplicate classes from
SettingsContentViewController that were left behind when
WorkspaceManagementView was originally extracted.
